+++ CHANGELOG.md
@@ -28,6 +28,8 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 * Added `about` to Polygon and Coords scale and rotate.
 
+* Add `to_geojson` to `PolygonList`
+
 ### Fixed
 * `kwimage.Detections` now correctly handles `None` data values. Previously
   `None` was converted to an `array(None, dtype=object)`.